layout, html: split paragraph.go and converter.go into per-seam files
paragraph.go and converter.go had grown into the two largest
hand-written files in the repo, mixing several concerns each. Move
functions into per-seam sibling files with zero behavior change:

- layout/paragraph_measure.go: word/line measurement (runMeasurer,
  MinWidth, MaxWidth, MeasureLines, MeasureHeight, measureWords, etc.)
- layout/paragraph_break.go: word breaking, hyphenation, and shaping
  (shapeAndMeasureWord, breakLongWords, hyphenateWord, splitWords, etc.)
- layout/paragraph_wrap.go: line building and wrapping (buildLine,
  wrapWords, linkSpans, truncateWithEllipsis)
- layout/paragraph_split.go: pagination (SplitAfterLine, PlanLayout,
  cloneWithWords, etc.)
- html/converter_assets.go: asset fetching, SSRF guarding, and byte
  budgeting (readAsset, resolveLocalAsset, DenyInternalHosts, etc.)
- html/converter_fonts.go: font-face loading and resolution
  (loadFontFaces, resolveFontForText, splitTextByFont, etc.)
- html/converter_dispatch.go: node walking and element dispatch
  (walkChildren, convertNode, convertElement, convertElementInner)

paragraph.go and converter.go now hold only construction/config,
Layout, and options/entrypoints. This extends the per-feature file
pattern already used elsewhere in the html package to layout.
